Feat/be/genre research

See merge request s08-bigdata-recom-sub2/S08P22D203!80

+++ server/src/main/java/com/emosaac/server/domain/user/Research.java
@@ -30,6 +30,6 @@ public class Research {
 
     private Long gerneId;
 
-    private Long type;
+    private int type;
 
 }
